-Ticket #281

+++ geodashboard-server/src/main/java-gen/stub/com/runwaysdk/geodashboard/MappableClass.java
@@ -95,10 +95,41 @@ public class MappableClass extends MappableClassBase implements com.runwaysdk.ge
   public void delete()
   {
     MdClass mdClass = this.getWrappedMdClass();
+    
+    /*
+     * Delete all layers which reference attributes on this type
+     */
+//    MetadataWrapperQuery query = new MetadataWrapperQuery(new QueryFactory());
+//    query.WHERE(query.getWrappedMdClass().EQ(mdClass));
+//    
+//    OIterator<? extends MetadataWrapper> iterator = query.getIterator();
+//
+//    try
+//    {
+//      while (iterator.hasNext())
+//      {
+//        MetadataWrapper wrapper = iterator.next();
+//        wrapper.delete();
+//      }
+//    }
+//    finally
+//    {
+//      iterator.close();
+//    }    
 
+    /*
+     * Delete all import mappings if they exist 
+     */
     TargetBinding binding = TargetBinding.getBinding(mdClass.definesType());
-    binding.delete();
 
+    if (binding != null)
+    {
+      binding.delete();
+    }
+
+    /*
+     * Delete all geo nodes
+     */
     OIterator<? extends GeoNode> nodes = this.getAllGeoNode();
 
     try
